Create the JSON object to read into the javascript visualization

This function creates the JSON object that feeds the visualization template for the LDAvis package visualization. It is a modified version of the one from the LDAvis package that does not use the document topics matrix to calculate the topic frequencies, instead the topic frequencies are passed as an argument. For a more detailed overview of LDAvis, see vignette("details", package = "LDAvis")

no_doc_topics_createJSON(phi = matrix(), topic.frequency,
  vocab = character(), term.frequency = integer(), R = 30,
  lambda.step = 0.01, mds.method = jsPCA, cluster,
  plot.opts = list(xlab = "PC1", ylab = "PC2"), reorder.topics = FALSE,
  ...)

Arguments

phi
matrix, with each row containing the distribution over terms for a topic, with as many rows as there are topics in the model, and as many columns as there are terms in the vocabulary.

topic.frequency
A numeric vector of the topic frequencies can be calculated with rcpp_parse_doc_topics.

vocab
character vector of the terms in the vocabulary (in the same order as the columns of phi). Each term must have at least one character.

term.frequency
integer vector containing the frequency of each term in the vocabulary.

R
integer, the number of terms to display in the barcharts of the interactive viz. Default is 30. Recommended to be roughly between 10 and 50.

lambda.step
a value between 0 and 1. Determines the interstep distance in the grid of lambda values over which to iterate when computing relevance. Default is 0.01. Recommended to be between 0.01 and 0.1.

mds.method
a function that takes phi as an input and outputs a K by 2 data.frame (or matrix). The output approximates the distance between topics. See jsPCA for details on the default method.

cluster
a cluster object created from the parallel package. If supplied, computations are performed using parLapply instead of lapply.

plot.opts
a named list used to customize various plot elements. By default, the x and y axes are labeled "PC1" and "PC2" (principal components 1 and 2), since jsPCA is the default scaling method.

reorder.topics
whether to re-order the K topics in order of decreasing proportion.

...
not currently used.

Value

A string containing JSON content which can be written to a file or feed into serVis for easy viewing/sharing. One element of this string is the new ordering of the topics.

Details

A modified version of LDAvis createJSON that doesn't compute the topic frequenices from doc topics. Instead it reads them in as an argument. The main computation is to loop through the topics and through the grid of lambda values (determined by lambda.step) to compute the R most relevant terms for each topic and value of lambda.

References

Sievert, C. and Shirley, K. (2014) LDAvis: A Method for Visualizing and Interpreting Topics, ACL Workshop on Interactive Language Learning, Visualization, and Interfaces. http://nlp.stanford.edu/events/illvi2014/papers/sievert-illvi2014.pdf