Rcpp Parse Document Topics File either from --output-doc-topics or from the parsed state file.

Allows for parsing information from the document terms matrix without loading the full thing into memory. Can accept as input either the doc topics file output from MALLET, or the gzipped doc topics generated by this package's rcpp_parse_topic_state function. From the document topic matrix file this function extracts the top documents for each topic. In addition, this function can optionally extract the top topics for each document, and if given as input a vector containing the lengths of each document, can extract the topic frequency for each topic.

rcpp_parse_doc_topics(fpath, topn, ttfile = "", q = 0.7,
  doc_lens = as.numeric(c()))

Arguments

fpath
A string specifying the path to either the --output-doc-topics file or a gzipped doc topics file generated by this package.

topn
A Number specifying the number of documents to save for each topic.

ttfile
**optional** A string specifying the path to an output gzipped file in which to save the top topics for each document.

q
**optional** A real number (defaults to 0.70) specifying the percentile of top topics to save for each document.

doc_lens
**optional** An integer vector containing the number of tokens in each document, if provided, the topic frequencies are calculated.

Value

A list with these components:

top_docs_inds

Numeric Matrix (K X topn) each cell has the index of the document starting at 0

top_docs_values

Numeric Matrix (K X topn) each cell has the doc topic score for the document, corresponds wiht top_docs_inds

topic_frequencies

A numeric vector of length(K) where each value is the topic proprotion, empty if doc_lens not provided